cleanup and stock chart component
