+++ src/components/analyser/details/Details.tsx
@@ -1,11 +1,10 @@
-// Based on Portfolio Details
+// Initally based on Portfolio Details, since then heavily modified
 import React from 'react';
 import { RouteComponentProps, useParams } from '@reach/router';
 import {
   LinearProgress,
   makeStyles,
-  Container,
-  useTheme,
+  Container
 } from '@material-ui/core';
 import { isAuthenticationError } from '../../../Errors';
 import * as API from '../../../analyser/APIClient';
@@ -13,15 +12,15 @@ import ErrorMessage from '../../shared/ErrorMessage';
 import DetailsHeader from './DetailsHeader';
 import DetailsOverview from './DetailsOverview';
 import KeyFigures from './KeyFigures';
-import StockChart from '../../shared/StockChart';
 import Dividends from './Dividends';
 import NewsComponent from './NewsComponent';
-import SectionDivider from './SectionDivider';
+import SectionDivider from '../../shared/SectionDivider';
 import Risks from './Risks';
 import BalanceSheetInfo from './BalanceSheetInfo';
 import Analysts from './Analysts';
 import AddToPortfolioButton from '../../shared/AddToPortfolioButton';
 import DetailsOverviewInfoBox from './DetailsOverviewInfoBox';
+import DetailsStockChart from './DetailsStockChart'
 
 // props type declaration
 export interface DetailsProps extends RouteComponentProps {
@@ -160,8 +159,6 @@ const Details: React.FC<DetailsProps> = ({ token, back }) => {
 
   const classes = useStyles();
 
-  const theme = useTheme();
-
   return (
     <>
       {!stockOverview ||
@@ -197,25 +194,19 @@ const Details: React.FC<DetailsProps> = ({ token, back }) => {
         newsList &&
         companyReports &&
         analystRecommendations && (
-          <div>
+          <>
             <DetailsHeader back={back} stock={stockOverview} />
             <Container className={classes.mainContent}>
               <DetailsOverview
                 stockOverview={stockOverview}
                 stockDetails={stockDetails}
               />
-              <StockChart
-                series={stockPerformance}
-                setPerformanceAll={setPerformanceAll}
-                axisColor={theme.palette.secondary.contrastText}
-                buttonBackgroundColor={theme.palette.primary.light}
-                buttonTextColor={theme.palette.primary.contrastText}
-                height={450}
-              />
+              <DetailsStockChart stockPerformance={stockPerformance} setPerformanceAll={setPerformanceAll}/> 
               <DetailsOverviewInfoBox
                 stockOverview={stockOverview}
                  stockDetails={stockDetails}
               />
+              <NewsComponent newsList={newsList} />
               <SectionDivider section="analyser.details.KeyFiguresHeader" />
               <KeyFigures />
               <Dividends />
@@ -225,10 +216,9 @@ const Details: React.FC<DetailsProps> = ({ token, back }) => {
                 overview={stockOverview}
               />
               <Risks />
-              <NewsComponent newsList={newsList} />
               <AddToPortfolioButton symbol={symbol} />
             </Container>
-          </div>
+          </>
         )}
     </>
   );
