+++ src/components/shared/SectionDivider.tsx
